// module5/cs_module5/src/app/model/customer/customer.ts
export class Customer {
  customerId?: number;
  customerName?: string;
  customerBirthday?: string;
  customerGender?: string;
  customerIdCard?: number;
  customerPhone?: number;
  customerEmail?: string;
  customerAddress?: string;
  customerType?: {
    customerTypeId?: number;
    customerTypeName?: string;
  };
}